With the addition of the intialize method on wrapped objects some of the customizations needed by OTDataDrawingToolView aren't needed anymore.
It would be nice to remove OTDataDrawingToolView all together, but it
is still needed to handle the creation of new objects when the tool
bar is used.  That could be factored out if the toolbar was setup in
the otml for the drawing, or it was setup by a service the drawing tool
looked for.


git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/data/datagraph@5816 6e01202a-0783-4428-890a-84243c50cc2b